ENH: CLI LocalCluster support

+++ ngff_zarr/to_multiscales.py
@@ -135,12 +135,15 @@ def _large_image_serialization(image: NgffImage, progress: Optional[RichDaskProg
     base_path = f"{image.name}-cache-{time.time()}"
 
     cache_store = config.cache_store
+    base_path_removed = False
     def remove_from_cache_store(sig_id, frame):
-        zarr.storage.rmdir(cache_store, base_path)
+        nonlocal base_path_removed
+        if not base_path_removed:
+            zarr.storage.rmdir(cache_store, base_path)
+            base_path_removed = True
     atexit.register(remove_from_cache_store, None, None)
     signal.signal(signal.SIGTERM, remove_from_cache_store)
     signal.signal(signal.SIGINT, remove_from_cache_store)
-    root = zarr.open_group(cache_store, mode='a')
 
     data = image.data
 
@@ -155,27 +158,30 @@ def _large_image_serialization(image: NgffImage, progress: Optional[RichDaskProg
         elif dim == 'c':
             rechunks[index] = 1
         else:
+            # rechunks[index] = min(optimized_chunks, data.shape[index], data.chunksize[index])
             rechunks[index] = min(optimized_chunks, data.shape[index])
 
     if progress:
-        progress.rich.log('Caching optimized chunks')
+        progress.rich.log('[blue]Caching optimized chunks')
 
-    data = data.rechunk(rechunks)
     if 'z' in dims:
         z_index = dims.index('z')
         slice_bytes = data.dtype.itemsize * data.shape[x_index] * data.shape[y_index]
+
         slab_slices = min(int(np.ceil(limit / slice_bytes)), data.shape[z_index])
         if optimized_chunks < data.shape[z_index]:
-            slab_slices = max(slab_slices, optimized_chunks)
+            slab_slices = min(slab_slices, optimized_chunks)
         split = _array_split(data, data.shape[z_index]//slab_slices, z_index)
 
+        rechunks[z_index] = slab_slices
+
+        data = data.rechunk(rechunks)
+
         split_arrays = []
-        split_paths = []
         for slab_index, slab in enumerate(split):
             path = base_path + f"/slab/{slab_index}"
-            path_group = root.create_group(path)
-            if progress:
-                progress.add_next_task(f"Caching z-slab {slab_index+1} of {len(split)}")
+            if isinstance(progress, RichDaskProgress):
+                progress.add_next_task(f"[blue]Caching z-slab {slab_index+1} of {len(split)}")
             arr = dask.array.to_zarr(
                 slab,
                 cache_store,
@@ -187,11 +193,11 @@ def _large_image_serialization(image: NgffImage, progress: Optional[RichDaskProg
             split_arrays.append(arr)
         data = dask.array.concatenate(split_arrays)
         if optimized_chunks < data.shape[z_index] and slab_slices < optimized_chunks:
+            rechunks[z_index] = optimized_chunks
             data = data.rechunk(rechunks)
             path = base_path + f"/optimized_chunks"
-            path_group = root.create_group(path)
-            if progress:
-                progress.add_next_task(f"Caching z rechunk")
+            if isinstance(progress, RichDaskProgress):
+                progress.add_next_task(f"[blue]Caching z rechunk")
             data = dask.array.to_zarr(
                 data,
                 cache_store,
@@ -201,11 +207,12 @@ def _large_image_serialization(image: NgffImage, progress: Optional[RichDaskProg
                 return_stored=True,
             )
     else:
+        data = data.rechunk(rechunks)
         # TODO: Do we need to split / concat very large 2D images
         path = base_path + f"/optimized_chunks"
-        path_group = root.create_group(path)
-        if progress:
-            progress.add_next_task(f"Caching optimized chunks")
+        # path_group = root.create_group(path)
+        if isinstance(progress, RichDaskProgress):
+            progress.add_next_task(f"[blue]Caching optimized chunks")
         data = dask.array.to_zarr(
             data,
             cache_store,
@@ -220,7 +227,7 @@ def _large_image_serialization(image: NgffImage, progress: Optional[RichDaskProg
 
 def to_multiscales(
     data: Union[NgffImage, ArrayLike, MutableMapping, str, ZarrArray],
-    scale_factors: Union[int, Sequence[Union[Dict[str, int], int]]] = 64,
+    scale_factors: Union[int, Sequence[Union[Dict[str, int], int]]] = 128,
     method: Optional[Methods] = None,
     chunks: Optional[
         Union[
@@ -265,7 +272,7 @@ def to_multiscales(
 
     # IPFS and visualization friendly default chunks
     if "z" in ngff_image.dims:
-        default_chunks = 64
+        default_chunks = 128
     else:
         default_chunks = 256
     default_chunks = {d: default_chunks for d in ngff_image.dims}
